added getCategories function and refactoring data models packages
