+++ app/src/main/java/com/identitymanager/database/firestore/callbacks/GetAccountCallback.java
@@ -1,6 +1,6 @@
 package com.identitymanager.database.firestore.callbacks;
 
-import com.identitymanager.models.Account;
+import com.identitymanager.models.data.Account;
 
 public interface GetAccountCallback {
