+++ app/src/main/java/com/identitymanager/database/firestore/callbacks/GetAccountsCallback.java
@@ -1,6 +1,6 @@
 package com.identitymanager.database.firestore.callbacks;
 
-import com.identitymanager.models.Account;
+import com.identitymanager.models.data.Account;
 
 import java.util.List;
