+++ app/src/main/java/com/identitymanager/models/data/Account.java
@@ -1,4 +1,4 @@
-package com.identitymanager.models;
+package com.identitymanager.models.data;
 
 public class Account {
