+++ app/src/main/java/com/identitymanager/models/data/Category.java
@@ -1,0 +1,23 @@
+package com.identitymanager.models.data;
+
+public class Category {
+
+    private String id;
+    private String label;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getLabel() {
+        return label;
+    }
+
+    public void setLabel(String label) {
+        this.label = label;
+    }
+}
